+++ Examples/TGReduxKitDemo/TGReduxKitDemo/TGReduxKitDemoApp.swift
@@ -1,17 +1,10 @@
-//
-//  TGReduxKitDemoApp.swift
-//  TGReduxKitDemo
-//
-//  Created by 小苹果 on 2026/1/14.
-//
-
 import SwiftUI
 
 @main
 struct TGReduxKitDemoApp: App {
     var body: some Scene {
         WindowGroup {
-            ShoppingAppView()
+            DemoRootView()
         }
     }
 }